refactor: Manage EmotionContainer views based on current state
